feat: replace original header with navigation and update headers to use title
