+++ frontend/src/components/Header.tsx
@@ -1,110 +1,37 @@
 'use client';
 
-import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-import { clsx } from 'clsx';
-import { Activity, BarChart3, List, Home, MessageSquare, Users, Menu, X, Bell } from 'lucide-react';
-import { useState } from 'react';
+import { Activity } from 'lucide-react';
 
-const navigation = [
-  { name: '대시보드', href: '/', icon: Home },
-  { name: '공지사항', href: '/notices', icon: Bell },
-  { name: '서비스 현황', href: '/services', icon: List },
-  { name: '통계', href: '/analytics', icon: BarChart3 },
-  { name: '정보 제보', href: '/report', icon: MessageSquare },
-  { name: '기여자', href: '/contributors', icon: Users },
-];
+interface HeaderProps {
+  title: string;
+  description?: string;
+  lastUpdated?: string | Date | null;
+}
 
-export default function Header() {
-  const pathname = usePathname();
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
+export default function Header({
+  title,
+  description,
+  lastUpdated,
+}: HeaderProps) {
+  const formattedTime = lastUpdated
+    ? new Date(lastUpdated).toLocaleString('ko-KR')
+    : '로딩 중...';
 
   return (
-    <header className="fixed top-0 left-0 right-0 z-50 bg-white shadow-sm border-b border-gray-200">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="flex justify-between items-center h-16">
-          {/* 로고 */}
-          <div className="flex items-center">
-            <Link href="/" className="flex items-center space-x-2">
-              <div className="w-8 h-8 bg-blue-600 rounded-lg flex items-center justify-center">
-                <Activity className="w-5 h-5 text-white" />
-              </div>
-              <span className="text-xl font-bold text-gray-900">
-                정부 사이트 장애 현황
-              </span>
-            </Link>
-          </div>
-
-          {/* 네비게이션 */}
-          <nav className="hidden md:flex space-x-8">
-            {navigation.map((item) => {
-              const Icon = item.icon;
-              const isActive = pathname === item.href;
-              
-              return (
-                <Link
-                  key={item.name}
-                  href={item.href}
-                  className={clsx(
-                    'flex items-center space-x-2 px-3 py-2 rounded-md text-sm font-medium transition-colors',
-                    isActive
-                      ? 'bg-blue-50 text-blue-700'
-                      : 'text-gray-600 hover:text-gray-900 hover:bg-gray-50'
-                  )}
-                >
-                  <Icon className="w-4 h-4" />
-                  <span>{item.name}</span>
-                </Link>
-              );
-            })}
-          </nav>
-
-          {/* 모바일 메뉴 버튼 */}
-          <div className="md:hidden">
-            <button
-              type="button"
-              onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-              className="text-gray-600 hover:text-gray-900 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 rounded-md p-2"
-            >
-              <span className="sr-only">{isMobileMenuOpen ? '메뉴 닫기' : '메뉴 열기'}</span>
-              {isMobileMenuOpen ? (
-                <X className="h-6 w-6" />
-              ) : (
-                <Menu className="h-6 w-6" />
-              )}
-            </button>
-          </div>
-        </div>
+    <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-6">
+      {/* 왼쪽: 제목 + 설명 */}
+      <div>
+        <h1 className="text-3xl font-bold text-gray-900">{title}</h1>
+        {description && (
+          <p className="mt-2 text-gray-600">{description}</p>
+        )}
       </div>
 
-      {/* 모바일 네비게이션 */}
-      {isMobileMenuOpen && (
-        <div className="md:hidden border-t border-gray-200">
-          <div className="px-2 pt-2 pb-3 space-y-1">
-            {navigation.map((item) => {
-              const Icon = item.icon;
-              const isActive = pathname === item.href;
-              
-              return (
-                <Link
-                  key={item.name}
-                  href={item.href}
-                  onClick={() => setIsMobileMenuOpen(false)}
-                  className={clsx(
-                    'flex items-center space-x-2 px-3 py-2 rounded-md text-base font-medium',
-                    isActive
-                      ? 'bg-blue-50 text-blue-700'
-                      : 'text-gray-600 hover:text-gray-900 hover:bg-gray-50'
-                  )}
-                >
-                  <Icon className="w-5 h-5" />
-                  <span>{item.name}</span>
-                </Link>
-              );
-            })}
-          </div>
-        </div>
-      )}
-    </header>
+      {/* 오른쪽: 업데이트 정보 */}
+      <div className="mt-4 sm:mt-0 flex items-center space-x-2 text-sm text-gray-500">
+        <Activity className="w-4 h-4 text-blue-500" />
+        <span>마지막 업데이트: {formattedTime}</span>
+      </div>
+    </div>
   );
 }
